🚸 Remove the difference between a blocking answer and a blocked answer
This is up for debate, but my personal view is that the user can see no difference between a blocking answer and a blocked answer. The difference is on a database level, and is an abstract concept to wrap your head around. Hence I have removed the visual style to destinguish these options