Add answer defined selections
If configured, the frontend will display an mutli-selection component allowing the user to add an extra layer for decision making, e. g. 'i want free licences' -> 'which licenses do you prefer?'. Each 'hit' on these extra layer values will result in an half point positive hit for result calculation. The Matched tags will be displayed at the result list per distribution.

This is preparation for:

#111
#184 (partly)
Also might be an idea for #149, but this requires more work as question structure need to be adapted based on user answers as this change is only to refine results