add security.txt
Hi @cmllr!

I added the security.txt file in the distrochooser project. Please, accept this pull-request and include the security.txt file in the distrochooser project.

The main purpose of security.txt is to help make things easier for companies and security researchers when trying to secure platforms. Thanks to security.txt, security researchers can easily get in touch with companies about security issues.

distrochooser is an open and online project (website) to facilitate the choice of linux distribution for ordinary users. In that sense, there's a good reason to include the security.txt file in case someone finds a security flaw on the site.